requirements.txt
streamlit
chromadb
sentence-transformers
transformers
torch
pdfplumber
python-dotenv
requests